# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Capabilities": {
        "description": "`GET /v1/capabilities`",
        "properties": {
          "backends": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "budgets": {
            "$ref": "#/components/schemas/Budgets"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceInfo"
            },
            "type": "array"
          },
          "execution_profile_schema_version": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "features": {
            "$ref": "#/components/schemas/Features"
          },
          "kernel_catalog_fingerprint": {
            "type": "string"
          },
          "rir_mode": {
            "description": "Effective process-wide mode: `off`, `observe`, `prefer`, or `require`.",
            "type": "string"
          },
          "rir_policy_latched": {
            "type": "boolean"
          },
          "unified_memory": {
            "description": "True when device and host memory are the same physical pool (Metal,\niGPU). Two independent budgets would then double-count the model.",
            "type": "boolean"
          }
        },
        "required": [
          "execution_profile_schema_version",
          "kernel_catalog_fingerprint",
          "rir_mode",
          "rir_policy_latched",
          "backends",
          "devices",
          "unified_memory",
          "budgets",
          "features"
        ],
        "type": "object"
      },
...
      "Environments": {
        "properties": {
          "environments": {
            "items": {
              "$ref": "#/components/schemas/EnvironmentEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "environments"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP control plane for LoRA training runs.",
    "title": "retrograd-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/capabilities": {
      "get": {
        "operationId": "get_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Devices, memory and effective budgets"
      }
    },
...
    "/v1/config-schema": {
      "get": {
        "operationId": "get_config_schema",
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "The configuration document as a JSON Schema 2020-12, with this server's marks"
      }
    },
    "/v1/datasets": {
      "get": {
        "operationId": "get_datasets",
        "parameters": [
          {
            "description": "Page size, 1 to 500. Default 50.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The id of the last item of the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "List stored datasets, newest first"
      },
      "post": {
        "operationId": "post_datasets",
        "parameters": [
          {
            "description": "A label for the dataset's own card.",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "auto (default) | text | jsonl | preference-jsonl. Absent or auto determines the format from the content, which never infers preference-jsonl: a file of pairs is declared.",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "application/x-ndjson": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "text/plain": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetView"
                }
              }
            },
            "description": "the same bytes were already stored; nothing was written twice"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetView"
                }
              }
            },
            "description": "created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Upload a dataset, content-addressed"
      }
    },
    "/v1/datasets/{id}": {
      "delete": {
        "operationId": "delete_datasets_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Forget a dataset"
      },
      "get": {
        "operationId": "get_datasets_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetView"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "One dataset's card"
      }
    },
    "/v1/datasets/{id}/preview": {
      "get": {
        "operationId": "get_datasets_id_preview",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "How many examples to show, 1 to 100. Default 5.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPreview"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "The first examples, exactly as read"
      }
    },
    "/v1/datasets/{id}/tokenize": {
      "post": {
        "operationId": "post_datasets_id_tokenize",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTokenization"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Measure real token lengths against a model's tokenizer"
      }
    },
    "/v1/defaults": {
      "get": {
        "operationId": "get_defaults",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Defaults"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Derivation rules and the settings applied by default"
      }
    },
    "/v1/environments": {
      "get": {
        "operationId": "get_environments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environments"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Environments the operator declared"
      }
    },
    "/v1/events": {
      "get": {
        "operationId": "get_events",
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Server-sent events across every run"
      }
    },
...
    "/v1/judges": {
      "get": {
        "operationId": "get_judges",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Judges"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Judges the operator declared"
      }
    },
    "/v1/mcp-servers": {
      "get": {
        "operationId": "get_mcp_servers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServers"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "MCP servers and the tools they really expose"
      }
    },
    "/v1/model-files": {
      "get": {
        "operationId": "get_model_files",
        "parameters": [
          {
            "description": "Read the disk again instead of the listing of the last 30 seconds.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelFileListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "GGUF files under the path roots, by name, for picking a model"
      }
    },
...
    "/v1/openapi.json": {
      "get": {
        "operationId": "get_openapi_json",
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "This document"
      }
    },
    "/v1/plan": {
      "post": {
        "operationId": "post_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Resolve a recipe without creating anything"
      }
    },
...
  }
}